organising the git directory
