--- TRAVEL.sql
@@ -1,137 +1,0 @@
-CREATE DATABASE TRAVEL;   
-USE TRAVEL;   
-
-CREATE TABLE  CUSTOMER ( CUST_ID INT IDENTITY PRIMARY KEY,   
-CUST_FName VARCHAR (20) NOT NULL,   
-CUST_LName VARCHAR (20) NOT NULL,   
-CUST_City VARCHAR (20) NOT NULL,   
-CUST_Phone VARCHAR (10) NOT NULL,   
-CUST_Email VARCHAR(50) NOT NULL,   
-CUST_LS VARCHAR (3) NOT NULL,   
-CUST_LOCALE VARCHAR (3) NOT NULL,   
-CUST_MS VARCHAR (3) NOT NULL,   
-);   
-INSERT INTO CUSTOMER  ( CUST_FName, CUST_LName, CUST_City, CUST_Phone, CUST_Email, CUST_LS, CUST_LOCALE, CUST_MS)  
-VALUES ('Jessie', 'Valenica', 'Dunedin', '0200939991','jessie56@gmail.com', 'FAM', 'MOU', 'DIN' ),  
-('Alex', 'Nash', 'Nelson', '0988562345','alex.nash98@hotmail.com', 'SIN', 'OTH', 'ALL'),  
-('Melody','Armstrong', 'Queenstown', '0218769822','melody@hotmail.com', 'FAM', 'OTH', 'BKF'),   
-('Eugene','Summer', 'Wellington', '0983452345', 'summer123@gmail.com', 'COU', 'SEA', 'ALL'),  
-('Tyrell', 'Harrington', 'Wellington', '0109121111', 'tyrell.harringtonoffical@gmail.com', 'COU', 'SEA', 'ALL');  
-
-
-CREATE TABLE RESORT (RESORT_ID INT IDENTITY PRIMARY KEY,   
-RESORT_Name VARCHAR(20) NOT NULL,   
-RESORT_CostPerNight NUMERIC (9,2) NOT NULL,   
-RESORT_LS VARCHAR (3) NOT NULL,   
-RESORT_LOCALE VARCHAR (3) NOT NULL,   
-RESORT_MS VARCHAR (3) NOT NULL,  
-);   
-INSERT INTO RESORT ( RESORT_Name, RESORT_CostPerNight, RESORT_LS, RESORT_LOCALE, RESORT_MS)  
-VALUES ('IVORY PARK', '900', 'FAM', 'OTH', 'ALL' ),  
-('MOUNTAIN VIEW', '700', 'COU', 'MOU', 'DIN'),  
-('SOUTHERN OCEANVIEW', '1000','FAM','SEA', 'BKF'),  
-('ACE', '3000', 'SIN','OTH','KIT'),  
-('EMBASSY', '500', 'COU','MOU', 'ALL');  
-
-CREATE TABLE ACTIVITES (ACT_ID INT IDENTITY PRIMARY KEY,   
-ACT_Name VARCHAR (20) NOT NULL,   
-ACT_Cost NUMERIC (7,2) NOT NULL,   
-);   
-INSERT INTO ACTIVITES (ACT_Name, ACT_Cost)  
-VALUES ('SKY DIVING','200'),  
-('ESCAPE ROOM', '60'),  
-('BOWLING', '30'),  
-('WINE TOURS', '100'),  
-('CANOEING', '80');  
-
-CREATE TABLE AGENT ( AGENT_ID INT IDENTITY PRIMARY KEY,   
-AGENT_Fname VARCHAR (20) NOT NULL,   
-AGENT_LName VARCHAR (20) NOT NULL,   
-AGENT_Phone CHAR (10) NOT NULL,   
-AGENT_Email VARCHAR (50) NOT NULL,   
-AGENT_Sales NUMERIC (9,2) NOT NULL   
-);  
-INSERT INTO AGENT (AGENT_Fname, AGENT_LName, AGENT_Phone, AGENT_Email, AGENT_Sales)  
-VALUES ('Niel','Williams','0946759222', 'niel@yahoo.com', '8000'),  
-('Grace', 'Hunter', '0391238765', 'gracehunter@gmail.com', '10000'),  
-('Karol', 'Watts', '8761239991', 'wattsorginal@gmail.com', '2000'),  
-('Rihanna', 'Wilson', '7645647777', 'rihanna1234@hotmail.com', '1900'),  
-('Autumn', 'Walkers', '2437653312', 'autumnwalkers65@gmail.com', '5000'); 
-
-CREATE TABLE VACATION (VACA_ID INT IDENTITY PRIMARY KEY,   
-VACA_StartDate DATE NOT NULL,   
-VACA_EndDate DATE NOT NULL,       
-CUST_ID INT NOT NULL FOREIGN KEY (CUST_ID) REFERENCES CUSTOMER(CUST_ID)
-ON DELETE CASCADE
-ON UPDATE CASCADE,
-RESORT_ID INT NOT NULL FOREIGN KEY (RESORT_ID) REFERENCES RESORT(RESORT_ID)
-ON DELETE CASCADE
-ON UPDATE CASCADE,
-AGENT_ID INT NOT NULL FOREIGN KEY (AGENT_ID) REFERENCES AGENT(AGENT_ID)
-ON DELETE CASCADE
-ON UPDATE CASCADE,
-);  
-
-INSERT INTO VACATION (VACA_StartDate, VACA_EndDate, CUST_ID, RESORT_ID, AGENT_ID) 
-VALUES 
-('2023-11-23','2023-11-30', 1, 2, 4), 
-('2023-09-10','2023-09-15', 3, 5, 2), 
-('2023-01-23','2023-01-31', 1, 1, 4), 
-('2023-12-01','2023-12-27', 4, 3, 3), 
-('2023-10-05','2023-10-10', 5, 2, 1), 
-('2023-03-27','2023-04-03', 2, 2, 2), 
-('2023-06-23','2023-06-27', 3, 4, 5), 
-('2023-02-28','2023-03-02', 5, 1, 2), 
-('2023-08-22','2023-08-26', 2, 4, 1), 
-('2023-05-05','2023-05-15', 4, 3, 5); 
-
-CREATE TABLE EXCURSION (EXC_ID INT IDENTITY PRIMARY KEY,   
-EXC_Date DATE NOT NULL,  
-ACT_ID INT NOT NULL FOREIGN KEY (ACT_ID) REFERENCES ACTIVITES (ACT_ID)  
-ON DELETE CASCADE
-ON UPDATE CASCADE,
-VACA_ID INT NOT NULL FOREIGN KEY (VACA_ID) REFERENCES VACATION (VACA_ID)
-ON DELETE CASCADE
-ON UPDATE CASCADE,
- );   
-
-INSERT INTO EXCURSION (EXC_Date, ACT_ID, VACA_ID)   
-VALUES  
-('2023-06-01', 2, 10),  
-('2023-06-07', 5, 4),  
-('2023-06-14', 2, 6),  
-('2023-06-20', 1, 8),  
-('2023-06-24', 2, 9),  
-('2023-06-27', 3, 10),  
-('2023-07-01', 4, 5),  
-('2023-07-13', 5, 7),  
-('2023-07-15', 3, 3),  
-('2023-07-20', 2, 2);
-
-
-SELECT CUSTOMER.CUST_FName + ' ' + CUSTOMER.CUST_LName, RESORT.RESORT_Name, VACATION.VACA_StartDate, VACATION.VACA_EndDate 
-FROM CUSTOMER 
-JOIN VACATION 
-ON CUSTOMER.CUST_ID = VACATION.VACA_ID 
-JOIN RESORT 
-ON RESORT.RESORT_ID = VACATION.RESORT_ID 
-WHERE CUSTOMER.CUST_City='Wellington'; 
-
-SELECT TOP 1 VACATION.RESORT_ID, RESORT.RESORT_Name, COUNT (VACATION .RESORT_ID) AS MostDemandingResort 
-FROM VACATION 
-JOIN RESORT ON VACATION.RESORT_ID = RESORT.RESORT_ID 
-GROUP BY VACATION.RESORT_ID, RESORT.RESORT_Name
-
-
-SELECT CUSTOMER.CUST_FName, CUSTOMER.CUST_LName, COUNT (VACATION.VACA_ID) AS VacationCount, 
-SUM (DATEDIFF(DAY, VACA_StartDate, VACA_EndDate)) as DayCount
-FROM CUSTOMER, VACATION
-WHERE CUSTOMER.CUST_ID = VACATION.CUST_ID
-GROUP BY CUSTOMER.CUST_ID, CUST_FName, CUST_LName
- 
-UPDATE AGENT
-SET AGENT_Sales = AGENT_Sales*1.1
- 
-DELETE CUSTOMER
-WHERE CUSTOMER.CUST_City ='Wellington';
-
